adg: avoid strong child-parent references
adg_entity_set_parent() was referencing the parent, hence generating a
child-parent strong reference. This is bad because when using
adg_container_add() a strong reference was already (and correctly) added
to the child, so there is a reference loop between child and parent.
This is already handled at AdgContainer level by using weak references
on the child instead, so dropping the strong set_parent reference seems
to be the right (TM) thing to do.