Revert "cpml: protect class data from NULL"
This reverts commit ce8e75569adaddf8e378e13cfb997f74e812e856.

This is the first of a serie of reverts. The CPML library does not have
any error reporting infrastructure so it cannot notify the user of any
API misuse.

So, what to do after a cpml_primitive_offset(NULL, 1)? The current
behavior of silently returning without notice is bad because can
propagate sublte errors in the above layers.

Not having other means to report, it is way better to crash immediately,
so the caller can fix its code. And this is what this reverts are trying
to restore.