cpml: make CpmlPrimitiveType an alias for cairo_path_data_type_t
Previously CpmlPrimitiveType was an enum on its own. This triggered any
kind of warning when using CpmlPrimitive on a cairo_path_data_type_t
variable and vice versa.

Removed the custom enum and used a combination of macro/typedef to
reach a similar effect but having both types interchangeable. Manually
defined the GObject wrapper to have the same type exported to
gobject-introspection.