cpml: drop CPML_GNUC_CONST macro
Actually the main function to get the primitive class from the type
number relies on a hash table, so further optimization are bogus.

Dropped the macro equivalent to G_GNUC_CONST because it is no more
needed (and because I suspect that macro was useless anyway).
Avoid relying on th