adg: fix type of va_start last parameter
Because the address of the last parameter may be used in the va_start()
macro, default argument promotions must not change the type of that
parameter. The compiler is free to implement enum as short int (or even
char), so this assumption will not be respected: short int and char
would be promoted to int. In this case the specification says the
behavior is undefined.

For some C99 reference regardling this topic see
http://stackoverflow.com/questions/24580503/error-when-pass-enum-in-a-function-with-variable-arguments

This was issue was highlighted by some strange failure of the test suite
with clang 3.9.0: https://travis-ci.org/ntd/adg/jobs/186622731