adg: add adg_dim_get_text() helper method
A new API to help the common rendering of nominal values.

Originally I wanted to leverage printf and friends but this ended up
being impossible: I must compile the argument list at run-time (because
of AdgDimStyle:number-arguments) but I cannot build va_list dynamically.

So this is a quite hackish implementation that relies on a regex that
looks for printf directives inside AdgDimStyle:number-format and
repeatedly calls g_ascii_formatd() [1].

[1] https://developer.gnome.org/glib/stable/glib-String-Utility-Functions.html#g-ascii-formatd