build: fix make distcheck for gobject-introspection
GObject introspection files should be installed in specific location not
dependent on $prefix. Automake makes this quite difficult.

A couple of additional configure options (--with-girdir and
--with-typelibdir) have been implemented for specifying custom locations
where the introspection files will be installed. Although not usually
needed, I use them in "make distcheck" to set a directory under $prefix
and make automake happy.