adg: add floating entity concept and API
When an entity is "floating", its extents does not concur on increasing
the extents of its own container. In other words, during the arrange
phase AdgContainer only considers the non-floating children to compute
its extents.

In particular, this affects how adg_canvas_autoscale() works: all
floating entities are not taken into consideration.

A typical example is the title block or any other annotation not
dependent from the model for positioning.