build: provide fallback to be able to build with old cairo releases
cairo_surface_set_device_scale() has been introduced in a relatively
recent cairo release (1.14.0) but this breaks Travic code integration
that has 1.13.1 installed.

Implemented a fallback function (a no-op really) to be able to build and
run the test suite but warns the user the media scaling feature will not
be supported.