demo: fix uninstalled check on autotools
When building with autotools, the real binary is put in a subdirectory
of `BUILDDIR` (`.libs`) and a wrapper file that bootstrap that
executable is created in its place. This breaks the uninstalled check
used since 4739eb85c6e5.

Check the prefix instead of the full string to solve this issue.